server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pyjwt==2.6.0
werkzeug==2.2.3 